Reorganize augur CLI
Replaced "augur run" with "augur server" family of commands.
Includes, start, stop, kill, and status commands.

Renamed augur configure generate to augur config init.
