+++ augur/cli/server.py
@@ -5,21 +5,28 @@ Augur library commands for controlling the backend components
 
 from copy import deepcopy
 import os, time, atexit, subprocess, click, atexit, logging, sys
+import psutil
+import signal
 import multiprocessing as mp
 import gunicorn.app.base
 from gunicorn.arbiter import Arbiter
 
+from augur.cli import initialize_logging, pass_config, pass_application
 from augur.housekeeper import Housekeeper
 from augur.server import Server, AUGUR_API_VERSION
-from augur.cli.util import stop_processes
 from augur.application import Application
+from augur.gunicorn import AugurGunicornApp
 
 logger = logging.getLogger("augur")
 
-@click.command("run")
+@click.group('server', short_help='Server commands')
+def cli():
+    pass
+
+@cli.command("start")
 @click.option("--disable-housekeeper", is_flag=True, default=False, help="Turns off the housekeeper")
 @click.option("--skip-cleanup", is_flag=True, default=False, help="Disables the old process cleanup that runs before Augur starts")
-def cli(disable_housekeeper, skip_cleanup):
+def start(disable_housekeeper, skip_cleanup):
     """
     Start Augur's backend server
     """
@@ -28,7 +35,7 @@ def cli(disable_housekeeper, skip_cleanup):
     logger.info(f"Using config file: {augur_app.config.config_file_location}")
     if not skip_cleanup:
         logger.debug("Cleaning up old Augur processes...")
-        stop_processes()
+        _broadcast_signal_to_processes()
         time.sleep(2)
     else:
         logger.debug("Skipping process cleanup")
@@ -40,6 +47,60 @@ def cli(disable_housekeeper, skip_cleanup):
     logger.info("Gunicorn server logs & errors will be written to logs/gunicorn.log")
     logger.info('Housekeeper update process logs will now take over.')
     Arbiter(master).run()
+
+@cli.command('stop')
+@initialize_logging
+def stop_server():
+    """
+    Sends SIGTERM to all Augur server & worker processes
+    """
+    _broadcast_signal_to_processes(attach_logger=True)
+
+
+@cli.command('kill')
+@initialize_logging
+def kill_server():
+    """
+    Sends SIGKILL to all Augur server & worker processes
+    """
+    _broadcast_signal_to_processes(signal=signal.SIGKILL, attach_logger=True)
+
+@cli.command('processes',)
+@initialize_logging
+def list_processes():
+    """
+    Outputs the name and process ID (PID) of all currently running backend Augur processes, including any workers. Will only work in a virtual environment.
+    """
+    processes = get_augur_processes()
+    for process in processes:
+        logger.info(f"Found process {process.pid}")
+
+def get_augur_processes():
+    processes = []
+    for process in psutil.process_iter(['cmdline', 'name', 'environ']):
+        if process.info['cmdline'] is not None and process.info['environ'] is not None:
+            try:
+                if os.getenv('VIRTUAL_ENV') in process.info['environ']['VIRTUAL_ENV'] and 'python' in ''.join(process.info['cmdline'][:]).lower():
+                    if process.pid != os.getpid():
+                        processes.append(process)
+            except KeyError:
+                pass
+    return processes
+
+def _broadcast_signal_to_processes(signal=signal.SIGTERM, attach_logger=False):
+    if attach_logger is True:
+        _logger = logging.getLogger("augur")
+    else:
+        _logger = logger
+    processes = get_augur_processes()
+    if processes != []:
+        for process in processes:
+            if process.pid != os.getpid():
+                logger.info(f"Stopping process {process.pid}")
+                try:
+                    process.send_signal(signal)
+                except psutil.NoSuchProcess as e:
+                    pass
 
 def initialize_components(augur_app, disable_housekeeper):
     master = None
@@ -83,59 +144,17 @@ def worker_start(worker_name=None, instance_number=0, worker_port=None):
 
 def exit(augur_app, worker_processes, master):
 
-        logger.info("Shutdown started for this Gunicorn worker...")
-        augur_app.shutdown()
+    logger.info("Shutdown started for this Gunicorn worker...")
+    augur_app.shutdown()
 
-        if worker_processes:
-            for process in worker_processes:
-                logger.debug("Shutting down worker process with pid: {}...".format(process.pid))
-                process.terminate()
+    if worker_processes:
+        for process in worker_processes:
+            logger.debug("Shutting down worker process with pid: {}...".format(process.pid))
+            process.terminate()
 
-        if master is not None:
-            logger.debug("Shutting down Gunicorn server")
-            master.halt()
+    if master is not None:
+        logger.debug("Shutting down Gunicorn server")
+        master.halt()
 
-        logger.info("Shutdown complete")
-        sys.exit(0)
-
-class AugurGunicornApp(gunicorn.app.base.BaseApplication):
-    """
-    Loads configurations, initializes Gunicorn, loads server
-    """
-
-    def __init__(self, options={}, augur_app=None):
-        self.options = options
-        self.augur_app = augur_app
-        self.manager = self.augur_app.manager
-        self.broker = self.augur_app.broker
-        self.housekeeper = self.augur_app.housekeeper
-        self.server = None
-        logger.debug(f"Gunicorn will start {self.options['workers']} worker processes")
-        super(AugurGunicornApp, self).__init__()
-
-    def load_config(self):
-        """
-        Sets the values for configurations
-        """
-        config = {key: value for key, value in self.options.items()
-                  if key in self.cfg.settings and value is not None}
-        for key, value in config.items():
-            self.cfg.set(key.lower(), value)
-
-    def get_augur_app(self):
-        """
-        Returns the loaded server
-        """
-        self.load()
-        return self.server.augur_app
-
-    def load(self):
-        """
-        Returns the loaded server
-        """
-        if self.server is None:
-            try:
-                self.server = Server(augur_app=self.augur_app)
-            except Exception as e:
-                logger.error(f"An error occured when Gunicorn tried to load the server: {e}")
-        return self.server.app
+    logger.info("Shutdown complete")
+    sys.exit(0)
