+++ scripts/control/start_augur.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
 echo "Starting backend server..."
-nohup augur run > /dev/null 2>&1 &
+nohup augur server start > /dev/null 2>&1 &
